refactor(library/Attempt): inlines `Attempt_Error_Actionable_from` within `Attempt_Error_Actionable`
