+++ src/library/Attempt/Error/Actionable.ts
@@ -2,13 +2,22 @@ import type {
   Branded,
 } from '@/library/typeUtilities';
 
+import type {
+  Attempt_Error_Interpreter,
+} from './Interpreter';
+
 import {
   Attempt_Error_NonActionable,
 } from './NonActionable';
 
 import {
-  Attempt_Error_Actionable_from,
-} from './assertActionable';
+  NonActionableBuiltInError,
+} from './NonActionableBuiltInError';
+
+import {
+  type AppropriatelyThrown,
+  PotentiallyActionable,
+} from './modifier';
 
 /** Extend this class to describe errors from which callers ought to recover */
 abstract class Attempt_Error_Actionable<
@@ -28,7 +37,29 @@ abstract class Attempt_Error_Actionable<
     });
   };
 
-  public static from = Attempt_Error_Actionable_from;
+  public static from<
+    SomeActionableError extends Attempt_Error_Actionable<string>,
+  >(
+    givenError: AppropriatelyThrown<Error>,
+    {
+      using: interpretationOf,
+    }: {
+      using: Attempt_Error_Interpreter<SomeActionableError>;
+    },
+  ): SomeActionableError {
+    const potentiallyActionableError = PotentiallyActionable.assume(givenError);
+    const definitelyActionableError = interpretationOf(potentiallyActionableError);
+
+    if (
+      definitelyActionableError !== null
+    ) return definitelyActionableError;
+
+    return Attempt_Error_NonActionable.rethrow(potentiallyActionableError, {
+      message: NonActionableBuiltInError.describes(givenError)
+        ? 'Neglected to prevent built-in error'
+        : 'Neglected to interpret or prevent potentially actionable error',
+    });
+  }
 }
 
 export {
